Set `sequenceSpace` on returned match
Value is equal to number of characters in sequence that matches
password (11 for digits, as '0' appears at each end of the sequence).

+++ Zxcvbn/DBSequenceMatch.h
@@ -11,7 +11,7 @@
 @interface DBSequenceMatch : DBMatch
 
 @property (strong, nonatomic) NSString *sequenceName;
-@property (nonatomic, assign) int sequenceSpace;
+@property (nonatomic, assign) NSUInteger sequenceSpace;
 @property (nonatomic, assign) BOOL ascending;
 
 @end
